refactor(parser): move inline link parsing to its own module
`inline_helpers.rs` crossed the repository's 350-line limit. Link parsing
is the natural seam: it is the one inline construct that has to parse its
own text before it can decide what it is, and it now carries the probe
cache that makes that affordable.
